+++ example/kotlin/composeApp/src/commonMain/kotlin/com/mauikmpexample/kotlin/shared/TestTopLevel.foo.kt
@@ -1,0 +1,10 @@
+package com.mauikmpexample.kotlin.shared
+
+import de.voize.mauikmp.annotation.MauiBinding
+
+@MauiBinding
+fun foo(
+    string: String,
+): String {
+    return "Hello $string"
+}